% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{integrate_p_times_fpld}
\alias{integrate_p_times_fpld}
\title{This is simply used as a test that integrate_p_times_fpld_simple works as it should}
\usage{
integrate_p_times_fpld(par, lower = 0, upper = 1)
}
\description{
This is simply used as a test that integrate_p_times_fpld_simple works as it should
}
