% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{locate_table_in_eklima_data}
\alias{locate_table_in_eklima_data}
\title{Function for finding the start and end of different types of
data tables in files from eklima}
\usage{
locate_table_in_eklima_data(path, start_text)
}
\description{
Function for finding the start and end of different types of
data tables in files from eklima
}
