% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-preprocessing.R
\name{num_vars_in_eklima_data}
\alias{num_vars_in_eklima_data}
\title{Find the number of variables in an eklima-file}
\usage{
num_vars_in_eklima_data(filepath, start)
}
\description{
Find the number of variables in an eklima-file
}
